@@ -6,49 +6,53 @@
   "author": "mrmin123 <minyoung.choi@gmail.com>",
   "license": "GPL-3.0",
   "scripts": {
-    "build": "webpack --env.prod --env.NODE_ENV=production --config ./docs-src/webpack.config.js",
-    "dev": "webpack --env.dev --env.NODE_ENV=development --progress --watch --config ./docs-src/webpack.config.js",
-    "dev-server": "webpack-dev-server --env.dev --env.NODE_ENV=development --hot --config ./docs-src/webpack.config.js --content-base docs/"
+    "build": "webpack --mode=production --env.NODE_ENV=production --config ./docs-src/webpack.config.js",
+    "dev": "webpack --mode=development --env.NODE_ENV=development --progress --watch --config ./docs-src/webpack.config.js",
+    "dev-server": "webpack-dev-server --mode=development --env.NODE_ENV=development --hot --config ./docs-src/webpack.config.js --content-base docs/"
   },
   "devDependencies": {
-    "babel-eslint": "^8.0.1",
-    "eslint": "^4.9.0",
-    "eslint-config-airbnb": "^16.1.0",
-    "eslint-plugin-babel": "^4.1.2",
-    "eslint-plugin-import": "^2.8.0",
-    "eslint-plugin-promise": "^3.6.0",
-    "eslint-plugin-react": "^7.4.0",
+    "babel-eslint": "^10.0.1",
+    "eslint": "^5.9.0",
+    "eslint-config-airbnb": "^17.1.0",
+    "eslint-plugin-babel": "^5.2.1",
+    "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-promise": "^4.0.1",
+    "eslint-plugin-react": "^7.11.1",
     "eslint-plugin-security": "^1.4.0",
-    "webpack-dev-server": "^2.9.3"
+    "webpack-cli": "^3.1.2",
+    "webpack-dev-server": "^3.1.10"
   },
   "dependencies": {
-    "babel-core": "^6.26.0",
-    "babel-loader": "^7.1.2",
+    "@babel/core": "^7.1.5",
+    "@babel/plugin-proposal-class-properties": "^7.1.0",
+    "@babel/preset-env": "^7.1.5",
+    "@babel/preset-react": "^7.0.0",
+    "@date-io/date-fns": "^0.0.2",
+    "@material-ui/core": "^3.5.1",
+    "babel-loader": "^8.0.4",
     "babel-polyfill": "^6.26.0",
-    "babel-preset-env": "^1.6.1",
-    "babel-preset-react": "^6.24.1",
     "babel-preset-stage-0": "^6.24.1",
-    "css-loader": "^0.28.7",
-    "eslint-plugin-jsx-a11y": "^6.0.2",
-    "file-loader": "^1.1.5",
-    "material-ui": "^1.0.0-beta.34",
-    "material-ui-time-picker": "^0.1.5",
-    "mdi": "^2.0.46",
-    "mdi-material-ui": "^2.0.1",
+    "css-loader": "^1.0.1",
+    "date-fns": "2.0.0-alpha.16",
+    "eslint-plugin-jsx-a11y": "^6.1.2",
+    "file-loader": "^2.0.0",
+    "material-ui-pickers": "^2.0.1",
+    "mdi": "^2.2.43",
+    "mdi-material-ui": "^5.6.0",
     "path": "^0.12.7",
-    "prop-types": "^15.6.0",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
-    "react-dropzone": "^4.2.3",
-    "react-redux": "^5.0.6",
-    "react-select": "^1.0.0-rc.10",
-    "redux": "^3.7.2",
+    "prop-types": "^15.6.2",
+    "react": "^16.6.3",
+    "react-dom": "^16.6.3",
+    "react-dropzone": "^7.0.1",
+    "react-redux": "^5.1.1",
+    "react-select": "^2.1.1",
+    "redux": "^4.0.1",
     "redux-logger": "^3.0.6",
-    "redux-thunk": "^2.2.0",
-    "resolve-url-loader": "^2.1.1",
+    "redux-thunk": "^2.3.0",
+    "resolve-url-loader": "^3.0.0",
     "save-as": "^0.1.8",
-    "style-loader": "^0.19.0",
-    "url-loader": "^0.6.2",
-    "webpack": "^3.8.1"
+    "style-loader": "^0.23.1",
+    "url-loader": "^1.1.2",
+    "webpack": "^4.25.1"
   }
 }
