@@ -1,49 +1,61 @@
-import React, { PureComponent, Fragment } from 'react'
+import React, { PureComponent } from 'react'
 import PropTypes from 'prop-types'
-import { withStyles } from 'material-ui/styles'
+import { withStyles } from '@material-ui/core/styles'
 
 import Select from 'react-select'
-import Grid from 'material-ui/Grid'
-import Typography from 'material-ui/Typography'
-import Switch from 'material-ui/Switch'
-import { InputLabel } from 'material-ui/Input'
-import { FormControl } from 'material-ui/Form'
+import Grid from '@material-ui/core/Grid'
+import Typography from '@material-ui/core/Typography'
+import Switch from '@material-ui/core/Switch'
+import InputLabel from '@material-ui/core/InputLabel'
+import FormControl from '@material-ui/core/FormControl'
 
 import Localize from 'containers/LocalizeContainer'
 import { styles } from 'components/BodyConfigStyles'
 
+import { FLEET_PRESETS } from 'types/formOptions'
 
-const FLEET_PRESETS = ['1', '2', '3', '4', '5', '6', '7', '8', '9', '10', '11', '12'].map(value => (
-  { value, label: value }))
 
 class BodyConfigPvP extends PureComponent {
-  state = this.props.config
-
-  componentWillReceiveProps = (nextProps) => {
-    if (this.props.config !== nextProps.config) {
-      this.setState(nextProps.config)
+  componentDidUpdate = (prevProps) => {
+    const {
+      config,
+      combatEnabled,
+      combatDisablePvP,
+      combatDisablePvPFleet,
+      updateObject,
+    } = this.props
+    if (combatEnabled && !prevProps.combatEnabled) {
+      // disable PvP and/or clear PvP Fleets accordingly depending on Combat Fleet Mode - this is in case the user
+      // defines a PvP-disabling fleet mode, disabled the Combat module, enables PvP and reassigns PvP fleets, then
+      // re-enables the Combat module
+      const updatedPvPConfig = {}
+      if (combatDisablePvP) updatedPvPConfig.pvpEnabled = false
+      if (combatDisablePvPFleet) updatedPvPConfig.pvpFleet = null
+      if (updatedPvPConfig) updateObject(config, updatedPvPConfig)
     }
   }
 
   render = () => {
     const {
       classes,
-    } = this.props
-    const {
+      config,
       pvpEnabled,
       pvpFleet,
       combatEnabled,
+      combatDisablePvP,
       combatDisablePvPFleet,
-    } = this.state
+      updateSwitch,
+      updateSelect,
+    } = this.props
     return (
-      <Fragment>
-        <Typography variant='display1'>
+      <>
+        <Typography variant='h5'>
           <Localize field='bodyConfig.pvpHeader' />
           <Switch
             className={classes.switch}
             checked={pvpEnabled}
-            onChange={
-              (event, checked) => this.setState({ pvpEnabled: checked }, () => this.props.callback(this.state))} />
+            onChange={(event, checked) => updateSwitch(config, event, checked, 'pvpEnabled')}
+            disabled={combatEnabled && combatDisablePvP} />
         </Typography>
 
         <Grid container spacing={0}>
@@ -57,19 +69,18 @@ class BodyConfigPvP extends PureComponent {
                 <Localize field='bodyConfig.pvpFleet' />
               </InputLabel>
               <Select
+                isClearable
                 className={classes.reactSelect}
-                simpleValue={true}
                 name='pvpFleet'
                 value={pvpFleet}
                 options={FLEET_PRESETS}
-                onChange={value => this.setState({ pvpFleet: value }, () => this.props.callback(this.state))}
-                disabled={!pvpEnabled || (combatEnabled && combatDisablePvPFleet)}
-                fullWidth />
+                onChange={value => updateSelect(config, value, 'pvpFleet')}
+                isDisabled={!pvpEnabled || (combatEnabled && combatDisablePvPFleet)} />
               <span className={classes.helperText}><Localize field='bodyConfig.pvpFleetDesc' /></span>
             </FormControl>
           </Grid>
         </Grid>
-      </Fragment>
+      </>
     )
   }
 }
@@ -77,7 +88,14 @@ class BodyConfigPvP extends PureComponent {
 BodyConfigPvP.propTypes = {
   classes: PropTypes.object.isRequired,
   config: PropTypes.object.isRequired,
-  callback: PropTypes.func.isRequired,
+  pvpEnabled: PropTypes.bool.isRequired,
+  pvpFleet: PropTypes.object,
+  combatEnabled: PropTypes.bool.isRequired,
+  combatDisablePvP: PropTypes.bool,
+  combatDisablePvPFleet: PropTypes.bool,
+  updateSwitch: PropTypes.func.isRequired,
+  updateSelect: PropTypes.func.isRequired,
+  updateObject: PropTypes.func.isRequired,
 }
 
 export default withStyles(styles)(BodyConfigPvP)
