@@ -6,10 +6,13 @@ import { createStore, applyMiddleware, compose } from 'redux'
 import thunk from 'redux-thunk'
 import { createLogger } from 'redux-logger'
 
-import Grid from 'material-ui/Grid'
-import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider'
-import { createMuiTheme } from 'material-ui/styles'
-import blueGrey from 'material-ui/colors/blueGrey'
+import Grid from '@material-ui/core/Grid'
+import MuiThemeProvider from '@material-ui/core/styles/MuiThemeProvider'
+import { createMuiTheme } from '@material-ui/core/styles'
+import blueGrey from '@material-ui/core/colors/blueGrey'
+
+import { MuiPickersUtilsProvider } from 'material-ui-pickers'
+import DateFnsUtils from '@date-io/date-fns'
 
 import { getUIMode } from 'actions/ui/mode'
 import { setUILanguage } from 'actions/ui/language'
@@ -20,6 +23,9 @@ import Footer from 'components/Footer'
 
 
 const theme = createMuiTheme({
+  typography: {
+    useNextVariants: true,
+  },
   palette: {
     primary: blueGrey,
     secondary: blueGrey,
@@ -30,9 +36,9 @@ const theme = createMuiTheme({
 
 let store = Reducers
 
-const reduxMiddleware = process.env.NODE_ENV === 'production' ?
-  applyMiddleware(thunk) :
-  applyMiddleware(thunk, createLogger())
+const reduxMiddleware = process.env.NODE_ENV === 'production'
+  ? applyMiddleware(thunk)
+  : applyMiddleware(thunk, createLogger())
 
 store = compose(reduxMiddleware)(createStore)(store)
 
@@ -42,11 +48,13 @@ store.dispatch(setUILanguage('en'))
 render(
   <Provider store={store}>
     <MuiThemeProvider theme={theme}>
-      <Grid container>
-        <MenuContainer />
-        <Body />
-        <Footer />
-      </Grid>
+      <MuiPickersUtilsProvider utils={DateFnsUtils}>
+        <Grid container>
+          <MenuContainer />
+          <Body />
+          <Footer />
+        </Grid>
+      </MuiPickersUtilsProvider>
     </MuiThemeProvider>
   </Provider>,
   document.getElementById('app')
