@@ -1,10 +1,10 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import { withStyles } from 'material-ui/styles'
+import { withStyles } from '@material-ui/core/styles'
 
-import Grid from 'material-ui/Grid'
-import Paper from 'material-ui/Paper'
-import Typography from 'material-ui/Typography'
+import Grid from '@material-ui/core/Grid'
+import Paper from '@material-ui/core/Paper'
+import Typography from '@material-ui/core/Typography'
 
 import Localize from 'containers/LocalizeContainer'
 
@@ -23,15 +23,15 @@ const BodyAbout = ({ classes }) => (
   <Grid container justify='center' spacing={0}>
     <Grid item xs={6}>
       <Paper className={classes.paper} elevation={0}>
-        <Typography variant='body1' className={classes.paragraph}>
+        <Typography variant='body2' className={classes.paragraph}>
           <Localize field='bodyAbout.intro1' />
         </Typography>
 
-        <Typography variant='body1' className={classes.paragraph}>
+        <Typography variant='body2' className={classes.paragraph}>
           <Localize field='bodyAbout.intro2' />
         </Typography>
 
-        <Typography variant='title'>
+        <Typography variant='h6'>
           <Localize field='bodyAbout.disclaimerHeader' />
         </Typography>
 
