@@ -1,224 +1,214 @@
-import React, { PureComponent, Fragment } from 'react'
+import React, { PureComponent } from 'react'
 import PropTypes from 'prop-types'
-import { withStyles } from 'material-ui/styles'
+import { withStyles } from '@material-ui/core/styles'
 
-import Select, { Creatable } from 'react-select'
-import Grid from 'material-ui/Grid'
-import Typography from 'material-ui/Typography'
-import Switch from 'material-ui/Switch'
-import Button from 'material-ui/Button'
-import { InputLabel } from 'material-ui/Input'
-import { FormControl, FormControlLabel } from 'material-ui/Form'
-import Checkbox from 'material-ui/Checkbox'
-import TimeInput from 'material-ui-time-picker'
+import Select from 'react-select'
+import CreatableSelect from 'react-select/lib/Creatable'
+import Grid from '@material-ui/core/Grid'
+import Typography from '@material-ui/core/Typography'
+import Switch from '@material-ui/core/Switch'
+import Button from '@material-ui/core/Button'
+import InputLabel from '@material-ui/core/InputLabel'
+import FormControl from '@material-ui/core/FormControl'
+import FormControlLabel from '@material-ui/core/FormControlLabel'
+import Checkbox from '@material-ui/core/Checkbox'
+import { TimePicker } from 'material-ui-pickers'
 import { ChevronRight } from 'mdi-material-ui'
 
 import Localize from 'containers/LocalizeContainer'
 import { styles } from 'components/BodyConfigStyles'
 
-
-const FLEET_PRESETS = ['1', '2', '3', '4', '5', '6', '7', '8', '9', '10', '11', '12'].map(value => (
-  { value, label: value }))
-const COMBAT_ENGINES = [
-  { value: 'live', label: <Localize field='bodyConfig.combatEngineLive' /> },
-  { value: 'legacy', label: <Localize field='bodyConfig.combatEngineLegacy' /> }]
-const MAPS = ['1-1', '1-2', '1-3', '1-4', '1-5', '1-6', '2-1', '2-2', '2-3', '2-4', '2-5', '3-1', '3-2', '3-3', '3-4',
-  '3-5', '4-1', '4-2', '4-3', '4-4', '4-5', '5-1', '5-2', '5-3', '5-4', '5-5', '6-1', '6-2', '6-3', '6-4', '6-5', '7-1',
-  'E-1', 'E-2', 'E-3', 'E-4', 'E-5', 'E-6', 'E-7', 'E-8']
-  .map(value => ({ value, label: value }))
-const COMBINED_FLEET_MODES = [
-  { value: '', label: <Localize field='bodyConfig.combatFleetModeStandard' /> },
-  { value: 'ctf', label: <Localize field='bodyConfig.combatFleetModeCTF' /> },
-  { value: 'stf', label: <Localize field='bodyConfig.combatFleetModeSTF' /> },
-  { value: 'transport', label: <Localize field='bodyConfig.combatFleetModeTransport' /> },
-  { value: 'striking', label: <Localize field='bodyConfig.combatFleetModeStriking' /> }]
-const COMBAT_NODE_COUNTS = ['1', '2', '3', '4', '5', '6', '7', '8', '9', '10'].map(value => (
-  { value, label: value }))
-const NODES = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ'.split('').map(value => ({ value, label: value }))
-NODES.push(...['Z1', 'Z2', 'Z3', 'Z4', 'Z5', 'Z6', 'Z7', 'Z8', 'Z9', 'ZZ1', 'ZZ2', 'ZZ3'].map(value => (
-  { value, label: value })))
-const FORMATIONS = [
-  { value: 'line_ahead', label: <Localize field='bodyConfig.combatFormationLineAhead' /> },
-  { value: 'double_line', label: <Localize field='bodyConfig.combatFormationDoubleLine' /> },
-  { value: 'diamond', label: <Localize field='bodyConfig.combatFormationDiamond' /> },
-  { value: 'echelon', label: <Localize field='bodyConfig.combatFormationEchelon' /> },
-  { value: 'line_abreast', label: <Localize field='bodyConfig.combatFormationLineAbreast' /> },
-  { value: 'vanguard', label: <Localize field='bodyConfig.combatFormationVanguard' /> },
-  { value: 'combinedfleet_1', label: <Localize field='bodyConfig.combatFormationCombinedFleet1' /> },
-  { value: 'combinedfleet_2', label: <Localize field='bodyConfig.combatFormationCombinedFleet2' /> },
-  { value: 'combinedfleet_3', label: <Localize field='bodyConfig.combatFormationCombinedFleet3' /> },
-  { value: 'combinedfleet_4', label: <Localize field='bodyConfig.combatFormationCombinedFleet4' /> }]
-const NIGHT_BATTLES = [
-  { value: 'True', label: <Localize field='bodyConfig.combatNightBattleTrue' /> },
-  { value: 'False', label: <Localize field='bodyConfig.combatNightBattleFalse' /> }]
-const DAMAGE_STATES = [
-  { value: 'heavy', label: <Localize field='bodyConfig.combatDamageStateHeavy' /> },
-  { value: 'moderate', label: <Localize field='bodyConfig.combatDamageStateModerate' /> },
-  { value: 'minor', label: <Localize field='bodyConfig.combatDamageStateMinor' /> }]
-const LBAS_GROUPS = ['1', '2', '3'].map(value => ({ value, label: value }))
+import {
+  FLEET_PRESETS, COMBAT_ENGINES, MAPS, COMBINED_FLEET_MODES, COMBAT_NODE_COUNTS, NODES, FORMATIONS, NIGHT_BATTLES,
+  DAMAGE_STATES, LBAS_GROUPS,
+} from 'types/formOptions'
 
 class BodyConfigCombat extends PureComponent {
-  state = this.props.config
-
-  componentWillReceiveProps = (nextProps) => {
-    if (this.props.config !== nextProps.config) {
-      this.setState(nextProps.config)
-    }
+  state = {
+    combatNodeSelect1: null,
+    combatNodeSelect2: null,
+    combatFormationsNode: null,
+    combatFormationsFormation: null,
+    combatNightBattlesNode: null,
+    combatNightBattlesMode: null,
   }
 
   handleCombatToggle = (event, checked) => {
     // when toggling the combat module, make sure to clear any conflicting expeditions as needed and disable the
-    // ship switcher as needed; these states are not actually used in here, but used to inform the other sections of
-    // available fields
+    // ship switcher as needed
+    const {
+      config,
+      combatFleetMode,
+      updateObject,
+    } = this.props
+
     if (checked) {
-      if (this.state.combatFleetMode === 'striking') {
-        this.setState({ combatEnabled: checked, expeditionsFleet3: [] }, () => this.props.callback(this.state))
-      } else if (['ctf', 'stf', 'transport'].includes(this.state.combatFleetMode)) {
-        this.setState({ combatEnabled: checked, expeditionsFleet2: [] }, () => this.props.callback(this.state))
+      if (combatFleetMode === 'striking') {
+        updateObject(config, { combatEnabled: checked, expeditionsFleet3: [] })
+      } else if (['ctf', 'stf', 'transport'].includes(combatFleetMode)) {
+        updateObject(config, { combatEnabled: checked, expeditionsFleet2: [] })
       } else {
-        this.setState({ combatEnabled: checked }, () => this.props.callback(this.state))
+        updateObject(config, { combatEnabled: checked })
       }
     } else {
-      this.setState({ combatEnabled: checked, shipSwitcherEnabled: false }, () => this.props.callback(this.state))
+      updateObject(config, { combatEnabled: checked, shipSwitcherEnabled: false })
     }
   }
 
   handleMapChange = (value) => {
-    // when changing the map, if the map is not an event map, disable the event reset module; these states are not
-    // actually used in here, but used to inform the event reset section of available fields
+    // when changing the map, if the map is not an event map, disable the event reset module
+    const {
+      config,
+      updateObject,
+    } = this.props
+
     if (value[0] !== 'E') {
-      this.setState({ combatMap: value, eventResetEnabled: false }, () => this.props.callback(this.state))
+      updateObject(config, { combatMap: value, eventResetEnabled: false })
     } else {
-      this.setState({ combatMap: value }, () => this.props.callback(this.state))
+      updateObject(config, { combatMap: value })
     }
   }
 
   handleFleetModeChange = (value) => {
-    // when changing the fleet mode, make sure to disable and clear any conflicting expeditions as needed; these states
-    // are not actually used in here, but used to inform the expedition section of available fields
-    if (value === 'striking') {
-      this.setState(
-        {
-          combatFleets: [],
-          combatFleetMode: value,
-          combatDisableExpeditionsFleet2: false,
-          expeditionsFleet3: [],
-          combatDisableExpeditionsFleet3: true,
-          pvpFleet: null,
-          combatDisablePvPFleet: true,
-        },
-        () => this.props.callback(this.state)
-      )
-    } else if (['ctf', 'stf', 'transport'].indexOf(value) > -1) {
-      this.setState(
-        {
-          combatFleets: [],
-          combatFleetMode: value,
-          expeditionsFleet2: [],
-          combatDisableExpeditionsFleet2: true,
-          combatDisableExpeditionsFleet3: false,
-          pvpFleet: null,
-          combatDisablePvPFleet: true,
-        },
-        () => this.props.callback(this.state)
-      )
+    // when changing the fleet mode, make sure to disable and clear any conflicting expeditions as needed
+    const {
+      config,
+      updateObject,
+    } = this.props
+
+    if (value.value === 'striking') {
+      updateObject(config, {
+        combatFleets: [],
+        combatFleetMode: value,
+        combatDisableExpeditionsFleet2: false,
+        expeditionsFleet3: [],
+        combatDisableExpeditionsFleet3: true,
+        combatDisablePvP: false,
+        pvpFleet: null,
+        combatDisablePvPFleet: false,
+      })
+    } else if (['ctf', 'stf', 'transport'].includes(value.value)) {
+      updateObject(config, {
+        combatFleets: [],
+        combatFleetMode: value,
+        expeditionsFleet2: [],
+        combatDisableExpeditionsFleet2: true,
+        combatDisableExpeditionsFleet3: false,
+        combatDisablePvP: true,
+        pvpEnabled: false,
+        pvpFleet: null,
+        combatDisablePvPFleet: true,
+      })
     } else {
-      this.setState(
-        {
-          combatFleetMode: value,
-          combatDisableExpeditionsFleet2: false,
-          combatDisableExpeditionsFleet3: false,
-          combatDisablePvPFleet: false,
-        },
-        () => this.props.callback(this.state)
-      )
+      updateObject(config, {
+        combatFleetMode: value,
+        combatDisableExpeditionsFleet2: false,
+        combatDisableExpeditionsFleet3: false,
+        combatDisablePvP: false,
+        combatDisablePvPFleet: false,
+      })
     }
   }
 
   handleCombatRetreatNodesChange = (value) => {
     // only allow the most recently selected numeric combat node count value; the actual script only allows uses the
     // smallest integer value in this field, but this exists as a sanitation step for the user
-    const validCombatRetreatNodes = []
+    const {
+      config,
+      updateObject,
+    } = this.props
+    const newCombatRetreatNodes = []
+
     let lastCombatNodeCount = null
     // determine last-specified number
-    value.split(',').forEach((val) => {
-      lastCombatNodeCount = !Number.isNaN(parseInt(val, 10)) ? parseInt(val, 10) : lastCombatNodeCount
+    value.forEach((opt) => {
+      lastCombatNodeCount = !Number.isNaN(parseInt(opt.value, 10)) ? parseInt(opt.value, 10) : lastCombatNodeCount
     })
     // filter out every integer value other than the last specified one
-    value.split(',').forEach((val) => {
-      if (Number.isNaN(parseInt(val, 10)) || parseInt(val, 10) === lastCombatNodeCount) {
-        validCombatRetreatNodes.push(val)
+    value.forEach((opt) => {
+      if (Number.isNaN(parseInt(opt.value, 10)) || parseInt(opt.value, 10) === lastCombatNodeCount) {
+        newCombatRetreatNodes.push(opt)
       }
     })
-    this.setState({
-      combatRetreatNodes: validCombatRetreatNodes.length > 0 ? validCombatRetreatNodes.join(',') : null,
-    }, () => this.props.callback(this.state))
+
+    updateObject(config, { combatRetreatNodes: newCombatRetreatNodes })
   }
 
   handleCombatNodeSelectAdd = (node, targetNode) => {
     // automatically add a node select option based on the two previous helper fields; also checks against previously
     // entered values so that existing node selects for a node are overwritten
-    const tempCombatNodeSelects = this.state.combatNodeSelects ? this.state.combatNodeSelects : ''
-    const tempCombatNodeSelectsObj = this.optionsNodeSplitter(tempCombatNodeSelects, '>')
-    tempCombatNodeSelectsObj[node] = targetNode
-    const combatNodeSelects = Object.keys(tempCombatNodeSelectsObj).sort().map(key =>
-      `${key}>${tempCombatNodeSelectsObj[key]}`).join(',')
-    this.setState(
-      { combatNodeSelect1: null, combatNodeSelect2: null, combatNodeSelects },
-      () => this.props.callback(this.state)
+    const {
+      config,
+      combatNodeSelects,
+      updateObject,
+    } = this.props
+
+    const newVal = `${node.value}>${targetNode.value}`
+    const newOpt = { label: newVal, value: newVal }
+
+    this.setState({ combatNodeSelect1: null, combatNodeSelect2: null })
+    const newCombatNodeSelects = combatNodeSelects.filter(
+      opt => opt.value.split('>')[0] !== newVal.split('>')[0]
     )
+    newCombatNodeSelects.push(newOpt)
+    updateObject(config, { combatNodeSelects: newCombatNodeSelects })
   }
 
   handleCombatFormationAdd = (node, formation) => {
     // automatically add a custom formation selection based on the two previous helper fields; also checks against
     // previously entered values so that existing formations for a node are overwritten
-    const tempCombatFormations = this.state.combatFormations ? this.state.combatFormations : ''
-    const tempCombatFormationsObj = this.optionsNodeSplitter(tempCombatFormations, ':')
+    const {
+      config,
+      combatFormations,
+      updateObject,
+    } = this.props
 
-    // if no node is specified, find next node number to apply formation to
-    const targetNode = node || this.findMaxNumericNode(tempCombatFormationsObj) + 1
-    tempCombatFormationsObj[targetNode] = formation
-    const combatFormations = Object.keys(tempCombatFormationsObj).sort().map(key =>
-      `${key}:${tempCombatFormationsObj[key]}`).join(',')
-    this.setState(
-      { combatFormationsNode: null, combatFormationsFormation: null, combatFormations },
-      () => this.props.callback(this.state)
+    const newVal = `${node.value}:${formation.value}`
+    const newOpt = { label: newVal, value: newVal }
+
+    this.setState({ combatFormationsNode: null, combatFormationsFormation: null })
+    const newCombatFormations = combatFormations.filter(
+      opt => opt.value.split(':')[0] !== newVal.split(':')[0]
     )
+    newCombatFormations.push(newOpt)
+    updateObject(config, { combatFormations: newCombatFormations })
   }
 
   handleCombatNightBattleAdd = (node, nightBattle) => {
     // automatically add a custom night battle selection based on the two previous helper fields; also checks against
     // previously entered values so that existing night battle selections for a node are overwritten
-    const tempCombatNightBattles = this.state.combatNightBattles ? this.state.combatNightBattles : ''
-    const tempCombatNightBattlesObj = this.optionsNodeSplitter(tempCombatNightBattles, ':')
-    // if no node is specified, find next node number to apply night battle mode to
-    const targetNode = node || this.findMaxNumericNode(tempCombatNightBattlesObj) + 1
-    tempCombatNightBattlesObj[targetNode] = nightBattle
-    const combatNightBattles = Object.keys(tempCombatNightBattlesObj).sort().map(key =>
-      `${key}:${tempCombatNightBattlesObj[key]}`).join(',')
-    this.setState(
-      { combatNightBattlesNode: null, combatNightBattlesMode: null, combatNightBattles },
-      () => this.props.callback(this.state)
+    const {
+      config,
+      combatNightBattles,
+      updateObject,
+    } = this.props
+
+    const newVal = `${node.value}:${nightBattle.value}`
+    const newOpt = { label: newVal, value: newVal }
+
+    this.setState({ combatNightBattlesNode: null, combatNightBattlesMode: null })
+    const newCombatNightBattles = combatNightBattles.filter(
+      opt => opt.value.split(':')[0] !== newVal.split(':')[0]
     )
+    newCombatNightBattles.push(newOpt)
+    updateObject(config, { combatNightBattles: newCombatNightBattles })
   }
 
   handleLBASGroupSelect = (value) => {
     // clear the LBAS node selects as needed based on the LBAS group selections
-    if (!value.includes('1')) {
-      this.setState({ combatLBASGroup1Node1: null, combatLBASGroup1Node2: null }, () => this.props.callback(this.state))
-    }
-    if (!value.includes('2')) {
-      this.setState({ combatLBASGroup2Node1: null, combatLBASGroup2Node2: null }, () => this.props.callback(this.state))
-    }
-    if (!value.includes('3')) {
-      this.setState({ combatLBASGroup3Node1: null, combatLBASGroup3Node2: null }, () => this.props.callback(this.state))
-    }
-    this.setState({ combatLBASGroups: value })
-  }
+    const {
+      config,
+      updateObject,
+    } = this.props
 
-  handleMiscOptionCheck = (event, checked) => {
-    // handling of misc option checkboxes
-    this.setState({ [event.target.value]: checked }, () => this.props.callback(this.state))
+    const newLBASNodes = {}
+    LBAS_GROUPS.forEach((opt) => {
+      if (value.filter(val => val.value === opt.value).length === 0) {
+        newLBASNodes[`combatLBASGroup${opt.label}Node1`] = null
+        newLBASNodes[`combatLBASGroup${opt.label}Node2`] = null
+      }
+    })
+    updateObject(config, { combatLBASGroups: value, ...newLBASNodes })
   }
 
   optionsNodeSplitter = (rawOption, divider) => {
@@ -249,22 +239,15 @@ class BodyConfigCombat extends PureComponent {
   render = () => {
     const {
       classes,
-    } = this.props
-    const {
+      config,
       combatEnabled,
-      combatFleets,
       combatEngine,
+      combatFleets,
       combatMap,
       combatFleetMode,
       combatRetreatNodes,
-      combatNodeSelect1,
-      combatNodeSelect2,
       combatNodeSelects,
-      combatFormationsNode,
-      combatFormationsFormation,
       combatFormations,
-      combatNightBattlesNode,
-      combatNightBattlesMode,
       combatNightBattles,
       combatRetreatLimit,
       combatRepairLimit,
@@ -281,25 +264,29 @@ class BodyConfigCombat extends PureComponent {
       combatOptionPortCheck,
       combatOptionClearStop,
       combatOptionLastNodePush,
+      updateSwitch,
+      updateSelect,
+      updateTime,
+    } = this.props
+    const {
+      combatNodeSelect1,
+      combatNodeSelect2,
+      combatFormationsNode,
+      combatFormationsFormation,
+      combatNightBattlesNode,
+      combatNightBattlesMode,
     } = this.state
 
-    const combatNodeSelectOptions = combatNodeSelects ?
-      combatNodeSelects.split(',').map(value => ({ value, label: value })) :
-      []
-    const combatFormationOptions = combatFormations ?
-      combatFormations.split(',').map(value => ({ value, label: value })) :
-      []
-    const combatNightBattleOptions = combatNightBattles ?
-      combatNightBattles.split(',').map(value => ({ value, label: value })) :
-      []
-    const combatLBASGroupsArray = combatLBASGroups ? combatLBASGroups.split(',') : []
-    const combatLBASGroup1NodesDisabled = !combatEnabled || combatLBASGroupsArray.indexOf('1') < 0
-    const combatLBASGroup2NodesDisabled = !combatEnabled || combatLBASGroupsArray.indexOf('2') < 0
-    const combatLBASGroup3NodesDisabled = !combatEnabled || combatLBASGroupsArray.indexOf('3') < 0
+    const combatLBASGroup1NodesDisabled = !combatEnabled
+      || combatLBASGroups.filter(opt => opt.value === '1').length === 0
+    const combatLBASGroup2NodesDisabled = !combatEnabled
+      || combatLBASGroups.filter(opt => opt.value === '2').length === 0
+    const combatLBASGroup3NodesDisabled = !combatEnabled
+      || combatLBASGroups.filter(opt => opt.value === '3').length === 0
 
     return (
-      <Fragment>
-        <Typography variant='display1'>
+      <>
+        <Typography variant='h5'>
           <Localize field='bodyConfig.combatHeader' />
           <Switch
             className={classes.switch}
@@ -315,14 +302,11 @@ class BodyConfigCombat extends PureComponent {
               </InputLabel>
               <Select
                 className={classes.reactSelect}
-                simpleValue={true}
                 name='combatEngine'
                 value={combatEngine}
                 options={COMBAT_ENGINES}
-                onChange={value => this.setState({ combatEngine: value }, () => this.props.callback(this.state))}
-                disabled={!combatEnabled}
-                clearable={false}
-                fullWidth />
+                onChange={value => updateSelect(config, value, 'combatEngine')}
+                isDisabled={!combatEnabled} />
             </FormControl>
           </Grid>
         </Grid>
@@ -330,7 +314,8 @@ class BodyConfigCombat extends PureComponent {
         <Grid container spacing={0}>
           <Grid item xs={12} sm={12} className={classes.formGrid}>
             <FormControl
-              disabled={!combatEnabled || combatFleetMode !== ''}
+              disabled={!combatEnabled
+                || combatFleetMode !== COMBINED_FLEET_MODES.filter(opt => opt.value === '')[0]}
               margin='normal'
               fullWidth
             >
@@ -338,15 +323,14 @@ class BodyConfigCombat extends PureComponent {
                 <Localize field='bodyConfig.combatFleets' />
               </InputLabel>
               <Select
-                multi
+                isMulti
                 className={classes.reactSelect}
-                simpleValue={true}
                 name='combatFleets'
                 value={combatFleets}
                 options={FLEET_PRESETS}
-                onChange={value => this.setState({ combatFleets: value }, () => this.props.callback(this.state))}
-                disabled={!combatEnabled || combatFleetMode !== ''}
-                fullWidth />
+                onChange={value => updateSelect(config, value, 'combatFleets')}
+                isDisabled={!combatEnabled
+                  || combatFleetMode !== COMBINED_FLEET_MODES.filter(opt => opt.value === '')[0]} />
               <span className={classes.helperText}><Localize field='bodyConfig.combatFleetsDesc' /></span>
             </FormControl>
           </Grid>
@@ -360,14 +344,11 @@ class BodyConfigCombat extends PureComponent {
               </InputLabel>
               <Select
                 className={classes.reactSelect}
-                simpleValue={true}
                 name='combatMap'
                 value={combatMap}
                 options={MAPS}
                 onChange={this.handleMapChange}
-                disabled={!combatEnabled}
-                clearable={false}
-                fullWidth />
+                isDisabled={!combatEnabled} />
             </FormControl>
           </Grid>
           <Grid item xs={12} sm={4} className={classes.formGrid}>
@@ -377,14 +358,11 @@ class BodyConfigCombat extends PureComponent {
               </InputLabel>
               <Select
                 className={classes.reactSelect}
-                simpleValue={true}
                 name='combatFleetMode'
                 value={combatFleetMode}
                 options={COMBINED_FLEET_MODES}
                 onChange={this.handleFleetModeChange}
-                disabled={!combatEnabled}
-                clearable={false}
-                fullWidth />
+                isDisabled={!combatEnabled} />
             </FormControl>
           </Grid>
           <Grid item xs={12} sm={4} className={classes.formGrid}>
@@ -393,15 +371,14 @@ class BodyConfigCombat extends PureComponent {
                 <Localize field='bodyConfig.combatRetreatNodes' />
               </InputLabel>
               <Select
-                multi
+                isMulti
+                isClearable
                 className={classes.reactSelect}
-                simpleValue={true}
                 name='combatRetreatNodes'
                 value={combatRetreatNodes}
                 options={COMBAT_NODE_COUNTS.concat(NODES)}
                 onChange={this.handleCombatRetreatNodesChange}
-                disabled={!combatEnabled}
-                fullWidth />
+                isDisabled={!combatEnabled} />
             </FormControl>
           </Grid>
 
@@ -412,13 +389,11 @@ class BodyConfigCombat extends PureComponent {
               </InputLabel>
               <Select
                 className={classes.reactSelect}
-                simpleValue={true}
                 name='combatNodeSelect1'
                 value={combatNodeSelect1}
                 options={NODES}
                 onChange={value => this.setState({ combatNodeSelect1: value })}
-                disabled={!combatEnabled}
-                fullWidth />
+                isDisabled={!combatEnabled} />
             </FormControl>
           </Grid>
           <Grid item xs={4} sm={2} className={classes.formGrid}>
@@ -428,21 +403,19 @@ class BodyConfigCombat extends PureComponent {
               </InputLabel>
               <Select
                 className={classes.reactSelect}
-                simpleValue={true}
                 name='combatNodeSelect2'
                 value={combatNodeSelect2}
                 options={NODES}
                 onChange={value => this.setState({ combatNodeSelect2: value })}
-                disabled={!combatEnabled}
-                fullWidth />
+                isDisabled={!combatEnabled} />
             </FormControl>
           </Grid>
           <Grid item xs={4} sm={1} className={classes.formGridButton}>
             <Button
               size='small'
               color='primary'
-              disabled={!combatEnabled ||
-                (!combatNodeSelect1 || !combatNodeSelect2 || combatNodeSelect1 === combatNodeSelect2)}
+              disabled={!combatEnabled
+                || (!combatNodeSelect1 || !combatNodeSelect2 || combatNodeSelect1 === combatNodeSelect2)}
               onClick={() => this.handleCombatNodeSelectAdd(combatNodeSelect1, combatNodeSelect2)}
             >
               <Localize field='bodyConfig.combatNodeSelectAdd' />
@@ -454,16 +427,16 @@ class BodyConfigCombat extends PureComponent {
               <InputLabel htmlFor='combatNodeSelects' shrink={true} className={classes.reactSelectLabel}>
                 <Localize field='bodyConfig.combatNodeSelects' />
               </InputLabel>
-              <Creatable
-                multi
+              <CreatableSelect
+                isMulti
+                isClearable
+                components={{ DropdownIndicator: null }}
                 className={classes.reactSelect}
-                simpleValue={true}
                 name='combatNodeSelects'
+                menuIsOpen={false}
                 value={combatNodeSelects}
-                options={combatNodeSelectOptions}
-                onChange={value => this.setState({ combatNodeSelects: value }, () => this.props.callback(this.state))}
-                disabled={!combatEnabled}
-                fullWidth />
+                onChange={value => updateSelect(config, value, 'combatNodeSelects')}
+                isDisabled={!combatEnabled} />
             </FormControl>
           </Grid>
 
@@ -474,13 +447,11 @@ class BodyConfigCombat extends PureComponent {
               </InputLabel>
               <Select
                 className={classes.reactSelect}
-                simpleValue={true}
                 name='combatFormationsNode'
                 value={combatFormationsNode}
                 options={combatEngine === 'legacy' ? COMBAT_NODE_COUNTS : COMBAT_NODE_COUNTS.concat(NODES)}
                 onChange={value => this.setState({ combatFormationsNode: value })}
-                disabled={!combatEnabled}
-                fullWidth />
+                isDisabled={!combatEnabled} />
             </FormControl>
           </Grid>
           <Grid item xs={4} sm={3} className={classes.formGrid}>
@@ -490,14 +461,11 @@ class BodyConfigCombat extends PureComponent {
               </InputLabel>
               <Select
                 className={classes.reactSelect}
-                simpleValue={true}
                 name='combatFormationsFormation'
                 value={combatFormationsFormation}
                 options={FORMATIONS}
-                onChange={
-                  value => this.setState({ combatFormationsFormation: value })}
-                disabled={!combatEnabled}
-                fullWidth />
+                onChange={value => this.setState({ combatFormationsFormation: value })}
+                isDisabled={!combatEnabled} />
             </FormControl>
           </Grid>
           <Grid item xs={4} sm={1} className={classes.formGridButton}>
@@ -516,16 +484,16 @@ class BodyConfigCombat extends PureComponent {
               <InputLabel htmlFor='combatFormations' shrink={true} className={classes.reactSelectLabel}>
                 <Localize field='bodyConfig.combatCustomFormations' />
               </InputLabel>
-              <Creatable
-                multi
+              <CreatableSelect
+                isMulti
+                isClearable
+                components={{ DropdownIndicator: null }}
                 className={classes.reactSelect}
-                simpleValue={true}
                 name='combatFormations'
+                menuIsOpen={false}
                 value={combatFormations}
-                options={combatFormationOptions}
-                onChange={value => this.setState({ combatFormations: value }, () => this.props.callback(this.state))}
-                disabled={!combatEnabled}
-                fullWidth />
+                onChange={value => updateSelect(config, value, 'combatFormations')}
+                isDisabled={!combatEnabled} />
             </FormControl>
           </Grid>
 
@@ -536,13 +504,11 @@ class BodyConfigCombat extends PureComponent {
               </InputLabel>
               <Select
                 className={classes.reactSelect}
-                simpleValue={true}
                 name='combatNightBattlesNode'
                 value={combatNightBattlesNode}
                 options={combatEngine === 'legacy' ? COMBAT_NODE_COUNTS : COMBAT_NODE_COUNTS.concat(NODES)}
                 onChange={value => this.setState({ combatNightBattlesNode: value })}
-                disabled={!combatEnabled}
-                fullWidth />
+                isDisabled={!combatEnabled} />
             </FormControl>
           </Grid>
           <Grid item xs={4} sm={3} className={classes.formGrid}>
@@ -552,14 +518,11 @@ class BodyConfigCombat extends PureComponent {
               </InputLabel>
               <Select
                 className={classes.reactSelect}
-                simpleValue={true}
                 name='combatNightBattlesMode'
                 value={combatNightBattlesMode}
                 options={NIGHT_BATTLES}
-                onChange={
-                  value => this.setState({ combatNightBattlesMode: value })}
-                disabled={!combatEnabled}
-                fullWidth />
+                onChange={value => this.setState({ combatNightBattlesMode: value })}
+                isDisabled={!combatEnabled} />
             </FormControl>
           </Grid>
           <Grid item xs={4} sm={1} className={classes.formGridButton}>
@@ -578,16 +541,16 @@ class BodyConfigCombat extends PureComponent {
               <InputLabel htmlFor='combatNightBattles' shrink={true} className={classes.reactSelectLabel}>
                 <Localize field='bodyConfig.combatCustomNightBattles' />
               </InputLabel>
-              <Creatable
-                multi
+              <CreatableSelect
+                isMulti
+                isClearable
+                components={{ DropdownIndicator: null }}
                 className={classes.reactSelect}
-                simpleValue={true}
                 name='combatNightBattles'
+                menuIsOpen={false}
                 value={combatNightBattles}
-                options={combatNightBattleOptions}
-                onChange={value => this.setState({ combatNightBattles: value }, () => this.props.callback(this.state))}
-                disabled={!combatEnabled}
-                fullWidth />
+                onChange={value => updateSelect(config, value, 'combatNightBattles')}
+                isDisabled={!combatEnabled} />
             </FormControl>
           </Grid>
 
@@ -598,12 +561,11 @@ class BodyConfigCombat extends PureComponent {
               </InputLabel>
               <Select
                 className={classes.reactSelect}
-                simpleValue={true}
                 name='combatRetreatLimit'
                 value={combatRetreatLimit}
                 options={DAMAGE_STATES}
-                onChange={value => this.setState({ combatRetreatLimit: value }, () => this.props.callback(this.state))}
-                disabled={!combatEnabled}
+                onChange={value => updateSelect(config, value, 'combatRetreatLimit')}
+                isDisabled={!combatEnabled}
                 clearable={false}
                 fullWidth />
             </FormControl>
@@ -615,28 +577,25 @@ class BodyConfigCombat extends PureComponent {
               </InputLabel>
               <Select
                 className={classes.reactSelect}
-                simpleValue={true}
                 name='combatRepairLimit'
                 value={combatRepairLimit}
                 options={DAMAGE_STATES}
-                onChange={value => this.setState({ combatRepairLimit: value }, () => this.props.callback(this.state))}
-                disabled={!combatEnabled}
+                onChange={value => updateSelect(config, value, 'combatRepairLimit')}
+                isDisabled={!combatEnabled}
                 clearable={false}
                 fullWidth />
             </FormControl>
           </Grid>
           <Grid item xs={12} sm={4} className={classes.formGrid}>
             <FormControl disabled={!combatEnabled} fullWidth>
-              <InputLabel htmlFor='combatRepairTimeLimit'>
-                <Localize field='bodyConfig.combatRepairTimeLimit' />
-              </InputLabel>
-              <TimeInput
+              <TimePicker
+                autoOk
                 id='combatRepairTimeLimit'
-                mode='24h'
-                autoOk={true}
+                label={<Localize field='bodyConfig.combatRepairTimeLimit' />}
                 value={combatRepairTimeLimit}
-                onChange={time => this.setState({ combatRepairTimeLimit: time }, () => this.props.callback(this.state))}
-                fullWidth />
+                ampm={false}
+                onChange={time => updateTime(config, time, 'combatRepairTimeLimit')}
+                disabled={!combatEnabled} />
             </FormControl>
           </Grid>
 
@@ -646,15 +605,14 @@ class BodyConfigCombat extends PureComponent {
                 <Localize field='bodyConfig.combatLBASGroups' />
               </InputLabel>
               <Select
-                multi
+                isMulti
+                isClearable
                 className={classes.reactSelect}
-                simpleValue={true}
                 name='combatLBASGroups'
                 value={combatLBASGroups}
                 options={LBAS_GROUPS}
                 onChange={this.handleLBASGroupSelect}
-                disabled={!combatEnabled}
-                fullWidth />
+                isDisabled={!combatEnabled} />
             </FormControl>
           </Grid>
           <Grid item xs={12} sm={3} className={classes.formGrid}>
@@ -665,22 +623,18 @@ class BodyConfigCombat extends PureComponent {
               <div className={classes.flexReset}>
                 <Select
                   className={classes.reactSelectHalfWidth}
-                  simpleValue={true}
                   name='combatLBASGroup1Node1'
                   value={combatLBASGroup1Node1}
                   options={NODES}
-                  onChange={
-                    value => this.setState({ combatLBASGroup1Node1: value }, () => this.props.callback(this.state))}
-                  disabled={combatLBASGroup1NodesDisabled} />
+                  onChange={value => updateSelect(config, value, 'combatLBASGroup1Node1')}
+                  isDisabled={combatLBASGroup1NodesDisabled} />
                 <Select
                   className={classes.reactSelectHalfWidth}
-                  simpleValue={true}
                   name='combatLBASGroup1Node2'
                   value={combatLBASGroup1Node2}
                   options={NODES}
-                  onChange={
-                    value => this.setState({ combatLBASGroup1Node2: value }, () => this.props.callback(this.state))}
-                  disabled={combatLBASGroup1NodesDisabled} />
+                  onChange={value => updateSelect(config, value, 'combatLBASGroup1Node2')}
+                  isDisabled={combatLBASGroup1NodesDisabled} />
               </div>
             </FormControl>
           </Grid>
@@ -692,22 +646,18 @@ class BodyConfigCombat extends PureComponent {
               <div className={classes.flexReset}>
                 <Select
                   className={classes.reactSelectHalfWidth}
-                  simpleValue={true}
                   name='combatLBASGroup2Node1'
                   value={combatLBASGroup2Node1}
                   options={NODES}
-                  onChange={
-                    value => this.setState({ combatLBASGroup2Node1: value }, () => this.props.callback(this.state))}
-                  disabled={combatLBASGroup2NodesDisabled} />
+                  onChange={value => updateSelect(config, value, 'combatLBASGroup2Node1')}
+                  isDisabled={combatLBASGroup2NodesDisabled} />
                 <Select
                   className={classes.reactSelectHalfWidth}
-                  simpleValue={true}
                   name='combatLBASGroup2Node2'
                   value={combatLBASGroup2Node2}
                   options={NODES}
-                  onChange={
-                    value => this.setState({ combatLBASGroup2Node2: value }, () => this.props.callback(this.state))}
-                  disabled={combatLBASGroup2NodesDisabled} />
+                  onChange={value => updateSelect(config, value, 'combatLBASGroup2Node2')}
+                  isDisabled={combatLBASGroup2NodesDisabled} />
               </div>
             </FormControl>
           </Grid>
@@ -719,80 +669,76 @@ class BodyConfigCombat extends PureComponent {
               <div className={classes.flexReset}>
                 <Select
                   className={classes.reactSelectHalfWidth}
-                  simpleValue={true}
                   name='combatLBASGroup3Node1'
                   value={combatLBASGroup3Node1}
                   options={NODES}
-                  onChange={
-                    value => this.setState({ combatLBASGroup3Node1: value }, () => this.props.callback(this.state))}
-                  disabled={combatLBASGroup3NodesDisabled} />
+                  onChange={value => updateSelect(config, value, 'combatLBASGroup3Node1')}
+                  isDisabled={combatLBASGroup3NodesDisabled} />
                 <Select
                   className={classes.reactSelectHalfWidth}
-                  simpleValue={true}
                   name='combatLBASGroup3Node2'
                   value={combatLBASGroup3Node2}
                   options={NODES}
-                  onChange={
-                    value => this.setState({ combatLBASGroup3Node2: value }, () => this.props.callback(this.state))}
-                  disabled={combatLBASGroup3NodesDisabled} />
+                  onChange={value => updateSelect(config, value, 'combatLBASGroup3Node2')}
+                  isDisabled={combatLBASGroup3NodesDisabled} />
               </div>
             </FormControl>
           </Grid>
 
           <Grid item xs={12} sm={12} className={classes.formGrid}>
             <FormControlLabel
-              control={
+              control={(
                 <Checkbox
                   checked={combatOptionCheckFatigue}
-                  onChange={this.handleMiscOptionCheck}
+                  onChange={(event, checked) => updateSwitch(config, event, checked, 'combatOptionCheckFatigue')}
                   disabled={!combatEnabled}
-                  value='combatOptionCheckFatigue' />
-              }
+                  name='combatOptionCheckFatigue' />
+              )}
               label={<Localize field='bodyConfig.combatCheckFatigue' />}
               disabled={!combatEnabled} />
             <FormControlLabel
-              control={
+              control={(
                 <Checkbox
                   checked={combatOptionReserveDocks}
-                  onChange={this.handleMiscOptionCheck}
+                  onChange={(event, checked) => updateSwitch(config, event, checked, 'combatOptionReserveDocks')}
                   disabled={!combatEnabled}
-                  value='combatOptionReserveDocks' />
-              }
+                  name='combatOptionReserveDocks' />
+              )}
               label={<Localize field='bodyConfig.combatReserveDocks' />}
               disabled={!combatEnabled} />
             <FormControlLabel
-              control={
+              control={(
                 <Checkbox
                   checked={combatOptionPortCheck}
-                  onChange={this.handleMiscOptionCheck}
+                  onChange={(event, checked) => updateSwitch(config, event, checked, 'combatOptionPortCheck')}
                   disabled={!combatEnabled}
-                  value='combatOptionPortCheck' />
-              }
+                  name='combatOptionPortCheck' />
+              )}
               label={<Localize field='bodyConfig.combatPortCheck' />}
               disabled={!combatEnabled} />
             <FormControlLabel
-              control={
+              control={(
                 <Checkbox
                   checked={combatOptionClearStop}
-                  onChange={this.handleMiscOptionCheck}
+                  onChange={(event, checked) => updateSwitch(config, event, checked, 'combatOptionClearStop')}
                   disabled={!combatEnabled}
-                  value='combatOptionClearStop' />
-              }
+                  name='combatOptionClearStop' />
+              )}
               label={<Localize field='bodyConfig.combatClearStop' />}
               disabled={!combatEnabled} />
             <FormControlLabel
-              control={
+              control={(
                 <Checkbox
                   checked={combatOptionLastNodePush}
-                  onChange={this.handleMiscOptionCheck}
+                  onChange={(event, checked) => updateSwitch(config, event, checked, 'combatOptionLastNodePush')}
                   disabled={!combatEnabled}
-                  value='combatOptionLastNodePush' />
-              }
+                  name='combatOptionLastNodePush' />
+              )}
               label={<Localize field='bodyConfig.combatLastNodePush' />}
               disabled={!combatEnabled} />
           </Grid>
         </Grid>
-      </Fragment>
+      </>
     )
   }
 }
@@ -800,7 +746,34 @@ class BodyConfigCombat extends PureComponent {
 BodyConfigCombat.propTypes = {
   classes: PropTypes.object.isRequired,
   config: PropTypes.object.isRequired,
-  callback: PropTypes.func.isRequired,
+  combatEnabled: PropTypes.bool.isRequired,
+  combatEngine: PropTypes.object.isRequired,
+  combatFleets: PropTypes.array,
+  combatMap: PropTypes.object.isRequired,
+  combatFleetMode: PropTypes.object,
+  combatRetreatNodes: PropTypes.array,
+  combatNodeSelects: PropTypes.array,
+  combatFormations: PropTypes.array,
+  combatNightBattles: PropTypes.array,
+  combatRetreatLimit: PropTypes.object.isRequired,
+  combatRepairLimit: PropTypes.object.isRequired,
+  combatRepairTimeLimit: PropTypes.object.isRequired,
+  combatLBASGroups: PropTypes.array,
+  combatLBASGroup1Node1: PropTypes.object,
+  combatLBASGroup1Node2: PropTypes.object,
+  combatLBASGroup2Node1: PropTypes.object,
+  combatLBASGroup2Node2: PropTypes.object,
+  combatLBASGroup3Node1: PropTypes.object,
+  combatLBASGroup3Node2: PropTypes.object,
+  combatOptionCheckFatigue: PropTypes.bool,
+  combatOptionReserveDocks: PropTypes.bool,
+  combatOptionPortCheck: PropTypes.bool,
+  combatOptionClearStop: PropTypes.bool,
+  combatOptionLastNodePush: PropTypes.bool,
+  updateSwitch: PropTypes.func.isRequired,
+  updateSelect: PropTypes.func.isRequired,
+  updateTime: PropTypes.func.isRequired,
+  updateObject: PropTypes.func.isRequired,
 }
 
 export default withStyles(styles)(BodyConfigCombat)
