@@ -1,98 +1,86 @@
-import React, { PureComponent, Fragment } from 'react'
+import React from 'react'
 import PropTypes from 'prop-types'
-import { withStyles } from 'material-ui/styles'
+import { withStyles } from '@material-ui/core/styles'
 
 import Select from 'react-select'
-import Grid from 'material-ui/Grid'
-import Typography from 'material-ui/Typography'
-import TextField from 'material-ui/TextField'
-import { InputLabel } from 'material-ui/Input'
-import { FormControl } from 'material-ui/Form'
+import Grid from '@material-ui/core/Grid'
+import Typography from '@material-ui/core/Typography'
+import TextField from '@material-ui/core/TextField'
+import InputLabel from '@material-ui/core/InputLabel'
+import FormControl from '@material-ui/core/FormControl'
 
 import Localize from 'containers/LocalizeContainer'
 import { styles } from 'components/BodyConfigStyles'
 
+import { PAUSE_OPTIONS } from 'types/formOptions'
 
-const PAUSE_OPTIONS = [
-  { value: true, label: <Localize field='bodyConfig.generalPauseTrue' /> },
-  { value: false, label: <Localize field='bodyConfig.generalPauseFalse' /> },
-]
+const BodyConfigGeneral = (props) => {
+  const {
+    classes,
+    config,
+    generalProgram,
+    generalJSTOffset,
+    generalPause,
+    updateText,
+    updateSelect,
+  } = props
 
-class BodyConfigGeneral extends PureComponent {
-  state = this.props.config
+  return (
+    <>
+      <Typography variant='h5'><Localize field='bodyConfig.generalHeader' /></Typography>
 
-  componentWillReceiveProps = (nextProps) => {
-    if (this.props.config !== nextProps.config) {
-      this.setState(nextProps.config)
-    }
-  }
-
-  render = () => {
-    const {
-      classes,
-    } = this.props
-    const {
-      generalProgram,
-      generalJSTOffset,
-      generalPause,
-    } = this.state
-    return (
-      <Fragment>
-        <Typography variant='display1'><Localize field='bodyConfig.generalHeader' /></Typography>
-
-        <Grid container spacing={0}>
-          <Grid item xs={12} sm={4} className={classes.formGrid}>
-            <TextField
-              id='generalProgram'
-              label={<Localize field='bodyConfig.generalProgram' />}
-              value={generalProgram}
-              onChange={
-                event => this.setState({ generalProgram: event.target.value }, () => this.props.callback(this.state))}
-              helperText={<Localize field='bodyConfig.generalProgramDesc' />}
-              className={classes.formControl}
-              fullWidth
-              margin='normal' />
-          </Grid>
-          <Grid item xs={12} sm={4} className={classes.formGrid}>
-            <TextField
-              id='generalJSTOffset'
-              label={<Localize field='bodyConfig.generalJSTOffset' />}
-              value={generalJSTOffset}
-              onChange={
-                event => this.setState({ generalJSTOffset: event.target.value }, () => this.props.callback(this.state))}
-              helperText={<Localize field='bodyConfig.generalJSTOffsetDesc' />}
-              className={classes.formControl}
-              fullWidth
-              type='number'
-              margin='normal' />
-          </Grid>
-          <Grid item xs={12} sm={4} className={classes.formGrid}>
-            <FormControl margin='normal' fullWidth>
-              <InputLabel htmlFor='generalPause' shrink={true} className={classes.reactSelectLabel}>
-                <Localize field='bodyConfig.generalPause' />
-              </InputLabel>
-              <Select
-                className={classes.reactSelect}
-                simpleValue={true}
-                name='generalPause'
-                value={generalPause}
-                options={PAUSE_OPTIONS}
-                clearable={false}
-                onChange={
-                  value => this.setState({ generalPause: value }, () => this.props.callback(this.state))}
-                fullWidth />
-            </FormControl>
-          </Grid>
+      <Grid container spacing={0}>
+        <Grid item xs={12} sm={4} className={classes.formGrid}>
+          <TextField
+            id='generalProgram'
+            label={<Localize field='bodyConfig.generalProgram' />}
+            value={generalProgram}
+            onChange={event => updateText(config, event, 'generalProgram')}
+            helperText={<Localize field='bodyConfig.generalProgramDesc' />}
+            className={classes.formControl}
+            fullWidth
+            margin='normal' />
         </Grid>
-      </Fragment>
-    )
-  }
+        <Grid item xs={12} sm={4} className={classes.formGrid}>
+          <TextField
+            id='generalJSTOffset'
+            label={<Localize field='bodyConfig.generalJSTOffset' />}
+            value={generalJSTOffset}
+            onChange={event => updateText(config, event, 'generalJSTOffset')}
+            helperText={<Localize field='bodyConfig.generalJSTOffsetDesc' />}
+            className={classes.formControl}
+            fullWidth
+            type='number'
+            margin='normal' />
+        </Grid>
+        <Grid item xs={12} sm={4} className={classes.formGrid}>
+          <FormControl margin='normal' fullWidth>
+            <InputLabel htmlFor='generalPause' shrink={true} className={classes.reactSelectLabel}>
+              <Localize field='bodyConfig.generalPause' />
+            </InputLabel>
+            <Select
+              className={classes.reactSelect}
+              name='generalPause'
+              value={generalPause}
+              options={PAUSE_OPTIONS}
+              clearable={false}
+              onChange={value => updateSelect(config, value, 'generalPause')}
+              fullWidth />
+          </FormControl>
+        </Grid>
+      </Grid>
+    </>
+  )
 }
 
 BodyConfigGeneral.propTypes = {
   classes: PropTypes.object.isRequired,
   config: PropTypes.object.isRequired,
-  callback: PropTypes.func.isRequired,
+  generalProgram: PropTypes.string.isRequired,
+  generalJSTOffset: PropTypes.string.isRequired,
+  generalPause: PropTypes.object.isRequired,
+  updateText: PropTypes.func.isRequired,
+  updateSelect: PropTypes.func.isRequired,
 }
 
 export default withStyles(styles)(BodyConfigGeneral)
