@@ -1,126 +1,119 @@
-import React, { PureComponent, Fragment } from 'react'
+import React, { PureComponent } from 'react'
 import PropTypes from 'prop-types'
-import { withStyles } from 'material-ui/styles'
+import { withStyles } from '@material-ui/core/styles'
 
-import Grid from 'material-ui/Grid'
-import Typography from 'material-ui/Typography'
-import Switch from 'material-ui/Switch'
-import { FormControlLabel } from 'material-ui/Form'
-import Checkbox from 'material-ui/Checkbox'
+import Grid from '@material-ui/core/Grid'
+import Typography from '@material-ui/core/Typography'
+import Switch from '@material-ui/core/Switch'
+import FormControlLabel from '@material-ui/core/FormControlLabel'
+import Checkbox from '@material-ui/core/Checkbox'
 
 import Localize from 'containers/LocalizeContainer'
 import { styles } from 'components/BodyConfigStyles'
 
 
 class BodyConfigQuests extends PureComponent {
-  state = this.props.config
-
-  componentWillReceiveProps = (nextProps) => {
-    if (this.props.config !== nextProps.config) {
-      this.setState(nextProps.config)
-    }
-  }
-
-  handleQuestGroupCheck = (event, checked) => {
-    // handle the quest group checkboxes and disable the entire module when all options are unchecked
-    const tempState = { [event.target.value]: checked }
-    const questGroups = [
-      'questsQuestGroupsDaily', 'questsQuestGroupsWeekly', 'questsQuestGroupsMonthly', 'questsQuestGroupOther']
-    let enableQuests = false
-    // ascertained if all groups are now disabled
-    questGroups.forEach((group) => {
-      if (group === event.target.value) {
-        if (checked === true) {
-          enableQuests = true
-        }
-      } else {
-        enableQuests = enableQuests || this.state[group]
-      }
-    })
-    // if all groups are disabled, also disable the quests section
-    if (!enableQuests) {
-      tempState.questsEnabled = false
-    }
-    this.setState(tempState, () => this.props.callback(this.state))
-  }
-
-  render = () => {
+  componentDidUpdate = (prevProps) => {
     const {
-      classes,
-    } = this.props
-    const {
+      config,
       questsEnabled,
       questsQuestGroupsDaily,
       questsQuestGroupsWeekly,
       questsQuestGroupsMonthly,
       questsQuestGroupsOthers,
-    } = this.state
+      updateObject,
+    } = this.props
+    if (questsEnabled && prevProps.questsEnabled) {
+      // disable the Quests module if all Quest groups are disabled
+      if (!questsQuestGroupsDaily && !questsQuestGroupsWeekly && !questsQuestGroupsMonthly
+          && !questsQuestGroupsOthers) {
+        updateObject(config, { questsEnabled: false })
+      }
+    }
+  }
+
+  render = () => {
+    const {
+      classes,
+      config,
+      questsEnabled,
+      questsQuestGroupsDaily,
+      questsQuestGroupsWeekly,
+      questsQuestGroupsMonthly,
+      questsQuestGroupsOthers,
+      updateSwitch,
+      updateObject,
+    } = this.props
+
     const allGroupsDisabled = (
-      !questsQuestGroupsDaily && !questsQuestGroupsWeekly && !questsQuestGroupsMonthly && !questsQuestGroupsOthers)
+      !questsQuestGroupsDaily && !questsQuestGroupsWeekly && !questsQuestGroupsMonthly
+      && !questsQuestGroupsOthers)
     return (
-      <Fragment>
-        <Typography variant='display1'>
+      <>
+        <Typography variant='h5'>
           <Localize field='bodyConfig.questsHeader' />
           <Switch
             className={classes.switch}
             checked={questsEnabled}
             onChange={
-              (event, checked) => {
-                const newState = { questsEnabled: checked }
-                if (checked && allGroupsDisabled) {
-                  newState.questsQuestGroupsDaily = true
-                  newState.questsQuestGroupsWeekly = true
-                  newState.questsQuestGroupsMonthly = true
-                  newState.questsQuestGroupsOthers = true
-                }
-                this.setState(newState, () => this.props.callback(this.state))
-              }
+              (event, checked) => updateObject(
+                config,
+                checked && allGroupsDisabled
+                  ? {
+                    questsEnabled: checked,
+                    questsQuestGroupsDaily: true,
+                    questsQuestGroupsWeekly: true,
+                    questsQuestGroupsMonthly: true,
+                    questsQuestGroupsOthers: true,
+                  }
+                  : { questsEnabled: checked }
+              )
             } />
         </Typography>
 
         <Grid item xs={12} sm={12} className={classes.formGrid}>
           <FormControlLabel
-            control={
+            control={(
               <Checkbox
                 checked={questsQuestGroupsDaily}
-                onChange={this.handleQuestGroupCheck}
+                onChange={(event, checked) => updateSwitch(config, event, checked, 'questsQuestGroupsDaily')}
                 disabled={!questsEnabled}
                 value='questsQuestGroupsDaily' />
-            }
+            )}
             label={<Localize field='bodyConfig.questsQuestGroupsDaily' />}
             disabled={!questsEnabled} />
           <FormControlLabel
-            control={
+            control={(
               <Checkbox
                 checked={questsQuestGroupsWeekly}
-                onChange={this.handleQuestGroupCheck}
+                onChange={(event, checked) => updateSwitch(config, event, checked, 'questsQuestGroupsWeekly')}
                 disabled={!questsEnabled}
                 value='questsQuestGroupsWeekly' />
-            }
+            )}
             label={<Localize field='bodyConfig.questsQuestGroupsWeekly' />}
             disabled={!questsEnabled} />
           <FormControlLabel
-            control={
+            control={(
               <Checkbox
                 checked={questsQuestGroupsMonthly}
-                onChange={this.handleQuestGroupCheck}
+                onChange={(event, checked) => updateSwitch(config, event, checked, 'questsQuestGroupsMonthly')}
                 disabled={!questsEnabled}
                 value='questsQuestGroupsMonthly' />
-            }
+            )}
             label={<Localize field='bodyConfig.questsQuestGroupsMonthly' />}
             disabled={!questsEnabled} />
           <FormControlLabel
-            control={
+            control={(
               <Checkbox
                 checked={questsQuestGroupsOthers}
-                onChange={this.handleQuestGroupCheck}
+                onChange={(event, checked) => updateSwitch(config, event, checked, 'questsQuestGroupsOthers')}
                 disabled={!questsEnabled}
                 value='questsQuestGroupsOthers' />
-            }
+            )}
             label={<Localize field='bodyConfig.questsQuestGroupsOthers' />}
             disabled={!questsEnabled} />
         </Grid>
-      </Fragment>
+      </>
     )
   }
 }
@@ -128,7 +121,13 @@ class BodyConfigQuests extends PureComponent {
 BodyConfigQuests.propTypes = {
   classes: PropTypes.object.isRequired,
   config: PropTypes.object.isRequired,
-  callback: PropTypes.func.isRequired,
+  questsEnabled: PropTypes.bool.isRequired,
+  questsQuestGroupsDaily: PropTypes.bool,
+  questsQuestGroupsWeekly: PropTypes.bool,
+  questsQuestGroupsMonthly: PropTypes.bool,
+  questsQuestGroupsOthers: PropTypes.bool,
+  updateSwitch: PropTypes.func.isRequired,
+  updateObject: PropTypes.func.isRequired,
 }
 
 export default withStyles(styles)(BodyConfigQuests)
