@@ -133,9 +133,8 @@ export const TMP = {
     shipSwitcherModalOffsetStartStart: null,
     shipSwitcherModalOffsetStartEnd: null,
     shipSwitcherModalOffsetCount: null,
-    shipSwitcherModalClasses: null,
-    shipSwitcherModalShips: null,
-    shipSwitcherModalShipsDesc: null,
+    shipSwitcherModalAssets: null,
+    shipSwitcherModalAssetsDesc: null,
     shipSwitcherModalLevel: null,
     shipSwitcherModalLockedIgnore: null,
     shipSwitcherModalLockedYes: null,
@@ -157,8 +156,8 @@ export const TMP = {
     intro: null,
     sikuliPath: null,
     sikuliPathDesc: null,
-    kcautoKaiPath: null,
-    kcautoKaiPathDesc: null,
+    kcautoPath: null,
+    kcautoPathDesc: null,
     commandHeader: null,
     noCommandNotice: null,
   },
