@@ -126,9 +126,8 @@ export const en = {
     shipSwitcherModalOffsetStartStart: 'Start of list',
     shipSwitcherModalOffsetStartEnd: 'End of list',
     shipSwitcherModalOffsetCount: 'Offset count',
-    shipSwitcherModalClasses: 'Select Class...',
-    shipSwitcherModalShips: 'Select Ship...',
-    shipSwitcherModalShipsDesc: `Custom ship-specific assets should be specified here. Specify in format <class>_<ship>,
+    shipSwitcherModalAssets: 'Select Asset (Class or Ship)...',
+    shipSwitcherModalAssetsDesc: `Custom ship-specific assets should be specified here. Specify in format <class>_<ship>,
       and save asset as 'shiplist_list_<class>_<ship>.png'`,
     shipSwitcherModalLevel: 'Filter on level',
     shipSwitcherModalLockedIgnore: 'Ignore Locks',
@@ -152,8 +151,8 @@ export const en = {
       generate the command you need to run in your command prompt/cmd/terminal to start kcauto.`,
     sikuliPath: 'Sikuli Path',
     sikuliPathDesc: 'Full path to the folder/directory where Sikuli is installed (where sikulix.jar exists)',
-    kcautoKaiPath: 'kcauto Path',
-    kcautoKaiPathDesc: 'Full path to the folder/directory where kcauto is installed (where config.ini exists)',
+    kcautoPath: 'kcauto Path',
+    kcautoPathDesc: 'Full path to the folder/directory where kcauto is installed (where config.ini exists)',
     commandHeader: 'Command',
     noCommandNotice: 'Fill out the above two fields to generate the command.',
   },
